fix(core): update formatEntityName import after string-utils move
Rebasing onto craig.marker/breadcrumb-title-case relocated
formatEntityName from hooks/use-entity-name/ into utils/string-utils --
these two consumers still pointed at the deleted path.
