+++ javascript/packages/core/components/views/phase-entity-view/phase-entity-view.tsx
@@ -10,7 +10,7 @@ import { CircleExclamationMarkKind } from '#core/components/illustrations/circle
 import { PageHeader } from '#core/components/page-header/page-header';
 import { Signpost } from '#core/components/signpost/signpost';
 import { useStudioParams } from '#core/hooks/routing/use-studio-params/use-studio-params';
-import { formatEntityName } from '#core/hooks/use-entity-name/use-entity-name';
+import { formatEntityName } from '#core/utils/string-utils';
 import { EntityTable } from './entity-table';
 
 import type { Theme } from 'baseui/theme';
